# products/dilchat/docs/openapi/dilchat.openapi.yaml
openapi: 3.1.0
info:
  title: DilChat API
  version: "1.0.0"
  summary: Consumer relationship + Vedic-astrology API for DilChat (Ugence Labs).
  description: |
    Design contract for the DilChat backend. REST/JSON over HTTPS, FastAPI-generated.
    Authoritative canon: `DILCHAT_DECISION_LOG.md`. Prose companion: `DILCHAT_API_SPEC.md`.

    Invariants encoded here:
    - Short-lived ES256 JWT access token (`Authorization: Bearer`) + rotating opaque refresh token.
    - Privacy scopes PRIVATE_A/PRIVATE_B/SHARED resolved server-side; couple membership re-checked
      on every shared request; default deny.
    - Content moves private -> shared ONLY via explicit consent grants.
    - Every derived response carries a `Provenance` block; AI never recomputes astrology.
    - Cross-private reads return 404 (never 403) so the other partner's private data existence
      is never leaked.
  contact:
    name: DilChat Engineering
    url: https://dilchat.com
  license:
    name: Proprietary — Ugence Labs
servers:
  - url: https://api.dilchat.com/v1
    description: Production (India-first region, DEC-018)
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [identity]
      operationId: authRegister
      summary: Create an account
      security: []
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Account created
          headers: { Location: { schema: { type: string } } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /auth/login:
    post:
      tags: [identity]
      operationId: authLogin
      summary: Password login
      security: []
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /auth/refresh:
    post:
      tags: [identity]
      operationId: authRefresh
      summary: Rotate refresh token, mint new access token
      security: []
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: Rotated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401':
          description: Invalid/reused refresh (AUTH_REFRESH_INVALID / AUTH_REFRESH_REUSE)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /auth/logout:
    post:
      tags: [identity]
      operationId: authLogout
      summary: Revoke the current session
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      responses:
        '204': { description: Session revoked }
        '401': { $ref: '#/components/responses/Problem' }
  /me:
    get:
      tags: [users]
      operationId: getMe
      summary: Current user profile
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [users]
      operationId: patchMe
      summary: Update profile
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name: { type: string }
                locale: { type: string }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
    delete:
      tags: [users]
      operationId: deleteMe
      summary: Request account deletion (async -> deletion_pending)
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirm]
              properties: { confirm: { type: string } }
      responses:
        '202':
          description: Deletion pending
          headers: { Location: { schema: { type: string } } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
...
  /birth-profiles:
    post:
      tags: [birth_profiles]
      operationId: createBirthProfile
      summary: Create a birth profile (private to owner)
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BirthProfileInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BirthProfile' }
        '400': { $ref: '#/components/responses/Problem' }
  /birth-profiles/{id}:
    get:
      tags: [birth_profiles]
      operationId: getBirthProfile
      summary: Get a birth-profile version (owner only; 404 for others)
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      responses:
        '200':
          description: Birth profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BirthProfile' }
        '404':
          description: Not found or masked cross-private (never 403)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
    patch:
      tags: [birth_profiles]
      operationId: patchBirthProfile
      summary: Supersede with a new immutable version
      parameters:
        - { $ref: '#/components/parameters/PathId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BirthProfileInput' }
      responses:
        '201':
          description: New version created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BirthProfile' }
  /me/birth-profile:
    get:
      tags: [birth_profiles]
      operationId: getMyBirthProfile
      summary: Current (latest-version) birth profile
      responses:
        '200':
          description: Birth profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BirthProfile' }
        '404': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived ES256 JWT access token. Verify against /.well-known/jwks.json.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Client-supplied key for exactly-once semantics on effectful POSTs (Redis, 24h TTL).
      schema: { type: string, maxLength: 255 }
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Problem:
      description: Problem+json error envelope
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, trace_id]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          enum:
            - VALIDATION_INVALID
            - VALIDATION_MISSING_FIELD
            - VALIDATION_OUT_OF_RANGE
            - AUTH_TOKEN_MISSING
            - AUTH_TOKEN_INVALID
            - AUTH_TOKEN_EXPIRED
            - AUTH_REFRESH_INVALID
            - AUTH_REFRESH_REUSE
            - AUTH_OTP_INVALID
            - SCOPE_DENIED
            - COUPLE_NOT_ACTIVE
            - CONSENT_REQUIRED
            - DUAL_APPROVAL_REQUIRED
            - NOT_FOUND
            - CONFLICT
            - INVITATION_USED
            - INVITATION_EXPIRED
            - IDEMPOTENCY_KEY_REUSED
            - AI_VALIDATION_FAILED
            - RATE_LIMITED
            - EPHEMERIS_UNAVAILABLE
            - INTERNAL
        detail: { type: string }
        trace_id: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 10 }
        display_name: { type: string }
        locale: { type: string, default: en-IN }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in, refresh_token]
      properties:
        access_token: { type: string }
        token_type: { type: string, enum: [Bearer] }
        expires_in: { type: integer, description: Access-token lifetime in seconds (600). }
        refresh_token: { type: string, description: Opaque, rotating; presented only to /auth/refresh. }
        session: { $ref: '#/components/schemas/Session' }
    AuthResult:
      type: object
      required: [tokens]
      properties:
        user: { $ref: '#/components/schemas/User' }
        tokens: { $ref: '#/components/schemas/TokenResponse' }
    User:
      type: object
      required: [id, email]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        display_name: { type: string }
        locale: { type: string }
        created_at: { type: string, format: date-time }
    Session:
      type: object
      properties:
        id: { type: string }
        device: { type: string }
        ip_hint: { type: string }
        created_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
        revoked: { type: boolean }
    Job:
      type: object
      required: [job_id, status]
      properties:
        job_id: { type: string }
        kind: { type: string }
        status:
          type: string
          enum: [queued, running, succeeded, failed, deletion_pending]
        progress: { type: number, minimum: 0, maximum: 1 }
        grace_until: { type: string, format: date-time }
        result:
          type: object
          properties:
            download_url: { type: string }
            expires_at: { type: string, format: date-time }
            size_bytes: { type: integer }
        error: { $ref: '#/components/schemas/Problem' }
        created_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
    Place:
      type: object
      properties:
        query: { type: string }
        display: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        iana_timezone: { type: string }
        coordinates_encrypted: { type: boolean }
    BirthProfileInput:
      type: object
      required: [birth_date, place]
      properties:
        birth_date: { type: string, format: date }
        birth_time: { type: string, description: Local civil time HH:MM:SS; separate from derived UTC. }
        birth_time_known: { type: boolean, default: true }
        place: { $ref: '#/components/schemas/Place' }
    BirthProfile:
      type: object
      required: [id, version, owner_user_id, birth_date]
      properties:
        id: { type: string }
        version: { type: integer }
        owner_user_id: { type: string }
        birth_date: { type: string, format: date }
        birth_time: { type: string }
        birth_time_known: { type: boolean }
        place: { $ref: '#/components/schemas/Place' }
        created_at: { type: string, format: date-time }
...